requirements.txt
fastapi
uvicorn
python-frontmatter
markdown
pygments
pymdown-extensions
pyyaml
